Extract abstract content method provider
