+++ TokenProvider/ContentMethodProvider.php
@@ -12,12 +12,10 @@
 
 namespace Symfony\Cmf\Component\RoutingAuto\TokenProvider;
 
-use Symfony\Cmf\Component\RoutingAuto\TokenProviderInterface;
 use Symfony\Cmf\Bundle\CoreBundle\Slugifier\SlugifierInterface;
 use Symfony\Component\OptionsResolver\OptionsResolverInterface;
-use Symfony\Cmf\Component\RoutingAuto\UriContext;
 
-class ContentMethodProvider implements TokenProviderInterface
+class ContentMethodProvider extends BaseContentMethodProvider
 {
     protected $slugifier;
 
@@ -26,29 +24,11 @@ class ContentMethodProvider implements TokenProviderInterface
         $this->slugifier = $slugifier;
     }
 
-    protected function checkMethodExists($object, $method)
-    {
-        if (!method_exists($object, $method)) {
-            throw new \InvalidArgumentException(sprintf(
-                'Method "%s" does not exist on object "%s"',
-                $method,
-                get_class($object)
-            ));
-        }
-    }
-
     /**
      * {@inheritDoc}
      */
-    public function provideValue(UriContext $uriContext, $options)
+    protected function normalizeValue($value, UriContext $uriContext, $options)
     {
-        $object = $uriContext->getSubjectObject();
-        $method = $options['method'];
-
-        $this->checkMethodExists($object, $method);
-
-        $value = $object->$method();
-
         if ($options['slugify']) {
             $value = $this->slugifier->slugify($value);
         }
@@ -61,9 +41,7 @@ class ContentMethodProvider implements TokenProviderInterface
      */
     public function configureOptions(OptionsResolverInterface $optionsResolver)
     {
-        $optionsResolver->setRequired(array(
-            'method',
-        ));
+        parent::configureOptions($optionsResolver);
 
         $optionsResolver->setDefaults(array(
             'slugify' => true,
